Add minStep to control classes
The standard structure for control fields in the Normative Type
Specification has a required "minStep" field in addition to ones for
low and high values. The utility classes Control and PVControlFactory
in org.epics.pvdata.property had no support for the minStep field. This
has been added in the same fashion as low and high and matches the C++
implementation behaviour.